bin: move load_zone() to ZoneConfig::load()
