+++ bin/src/lib.rs
@@ -16,6 +16,7 @@ use std::{
     net::{AddrParseError, Ipv4Addr, Ipv6Addr},
     path::{Path, PathBuf},
     str::FromStr,
+    sync::Arc,
     time::Duration,
 };
 
@@ -26,19 +27,31 @@ use serde::{self, Deserialize, Deserializer};
 
 use hickory_proto::rr::Name;
 use hickory_proto::ProtoError;
-use hickory_server::authority::ZoneType;
+#[cfg(feature = "dnssec-ring")]
+use hickory_server::authority::DnssecAuthority;
 #[cfg(feature = "dnssec-ring")]
 use hickory_server::dnssec::NxProofKind;
+#[cfg(feature = "blocklist")]
+use hickory_server::store::blocklist::BlocklistAuthority;
 #[cfg(feature = "blocklist")]
 use hickory_server::store::blocklist::BlocklistConfig;
 use hickory_server::store::file::FileConfig;
 #[cfg(feature = "resolver")]
+use hickory_server::store::forwarder::ForwardAuthority;
+#[cfg(feature = "resolver")]
 use hickory_server::store::forwarder::ForwardConfig;
+#[cfg(feature = "recursor")]
+use hickory_server::store::recursor::RecursiveAuthority;
 #[cfg(feature = "recursor")]
 use hickory_server::store::recursor::RecursiveConfig;
 #[cfg(feature = "sqlite")]
-use hickory_server::store::sqlite::SqliteConfig;
-use hickory_server::ConfigError;
+use hickory_server::store::sqlite::{SqliteAuthority, SqliteConfig};
+use hickory_server::{
+    authority::{AuthorityObject, ZoneType},
+    store::file::FileAuthority,
+    ConfigError,
+};
+use tracing::{debug, info, warn};
 
 static DEFAULT_PATH: &str = "/var/named"; // TODO what about windows (do I care? ;)
 static DEFAULT_PORT: u16 = 53;
@@ -321,6 +334,161 @@ impl ZoneConfig {
         }
     }
 
+    #[cfg_attr(not(feature = "dnssec-ring"), allow(unused_mut, unused))]
+    #[warn(clippy::wildcard_enum_match_arm)] // make sure all cases are handled despite of non_exhaustive
+    pub async fn load(&self, zone_dir: &Path) -> Result<Vec<Arc<dyn AuthorityObject>>, String> {
+        debug!("loading zone with config: {self:#?}");
+
+        let zone_name = self
+            .zone()
+            .map_err(|err| format!("failed to read zone name: {err}"))?;
+        let zone_type = self.zone_type();
+
+        // load the zone and insert any configured authorities in the catalog.
+
+        let mut authorities: Vec<Arc<dyn AuthorityObject>> = vec![];
+
+        #[cfg(feature = "blocklist")]
+        let handle_blocklist_store = |config| {
+            let zone_name = zone_name.clone();
+
+            async move {
+                Result::<Arc<dyn AuthorityObject>, String>::Ok(Arc::new(
+                    BlocklistAuthority::try_from_config(
+                        zone_name.clone(),
+                        zone_type,
+                        config,
+                        Some(zone_dir),
+                    )
+                    .await?,
+                ))
+            }
+        };
+
+        match &self.zone_type_config {
+            ZoneTypeConfig::Primary(server_config) | ZoneTypeConfig::Secondary(server_config) => {
+                debug!(
+                    "loading authorities for {zone_name} with stores {:?}",
+                    server_config.stores
+                );
+
+                let zone_name_for_signer = zone_name.clone();
+                let is_axfr_allowed = server_config.is_axfr_allowed();
+
+                for store in &server_config.stores {
+                    let authority: Arc<dyn AuthorityObject> = match store {
+                        #[cfg(feature = "sqlite")]
+                        ServerStoreConfig::Sqlite(config) => {
+                            let mut authority = SqliteAuthority::try_from_config(
+                                zone_name.clone(),
+                                zone_type,
+                                is_axfr_allowed,
+                                server_config.is_dnssec_enabled(),
+                                Some(zone_dir),
+                                config,
+                                #[cfg(feature = "dnssec-ring")]
+                                server_config.nx_proof_kind.clone(),
+                            )
+                            .await?;
+
+                            load_keys(&mut authority, zone_name_for_signer.clone(), server_config)
+                                .await?;
+                            Arc::new(authority)
+                        }
+
+                        ServerStoreConfig::File(config) => {
+                            let mut authority = FileAuthority::try_from_config(
+                                zone_name.clone(),
+                                zone_type,
+                                is_axfr_allowed,
+                                Some(zone_dir),
+                                config,
+                                #[cfg(feature = "dnssec-ring")]
+                                server_config.nx_proof_kind.clone(),
+                            )?;
+
+                            load_keys(&mut authority, zone_name_for_signer.clone(), server_config)
+                                .await?;
+                            Arc::new(authority)
+                        }
+                        _ => return empty_stores_error(),
+                    };
+
+                    authorities.push(authority);
+                }
+            }
+            ZoneTypeConfig::Forward { stores } => {
+                debug!(
+                    "loading authorities for {zone_name} with stores {:?}",
+                    stores
+                );
+
+                #[cfg_attr(
+                    not(any(feature = "blocklist", feature = "resolver")),
+                    allow(clippy::never_loop)
+                )]
+                for store in stores {
+                    let authority = match store {
+                        #[cfg(feature = "blocklist")]
+                        ForwardStoreConfig::Blocklist(config) => {
+                            handle_blocklist_store(config).await?
+                        }
+                        #[cfg(feature = "resolver")]
+                        ForwardStoreConfig::Forward(config) => {
+                            let forwarder = ForwardAuthority::try_from_config(
+                                zone_name.clone(),
+                                zone_type,
+                                config,
+                            )?;
+
+                            Arc::new(forwarder)
+                        }
+                        _ => return empty_stores_error(),
+                    };
+
+                    authorities.push(authority);
+                }
+            }
+            ZoneTypeConfig::Hint { stores } => {
+                debug!(
+                    "loading authorities for {zone_name} with stores {:?}",
+                    stores
+                );
+
+                #[cfg_attr(
+                    not(any(feature = "blocklist", feature = "recursor")),
+                    allow(clippy::never_loop)
+                )]
+                for store in stores {
+                    let authority: Arc<dyn AuthorityObject> = match store {
+                        #[cfg(feature = "blocklist")]
+                        HintStoreConfig::Blocklist(ref config) => {
+                            handle_blocklist_store(config).await?
+                        }
+                        #[cfg(feature = "recursor")]
+                        HintStoreConfig::Recursor(config) => {
+                            let recursor = RecursiveAuthority::try_from_config(
+                                zone_name.clone(),
+                                zone_type,
+                                config,
+                                Some(zone_dir),
+                            )
+                            .await?;
+
+                            Arc::new(recursor)
+                        }
+                        _ => return empty_stores_error(),
+                    };
+
+                    authorities.push(authority);
+                }
+            }
+        }
+
+        info!("zone successfully loaded: {}", self.zone()?);
+        Ok(authorities)
+    }
+
     // TODO this is a little ugly for the parse, b/c there is no terminal char
     /// returns the name of the Zone, i.e. the `example.com` of `www.example.com.`
     pub fn zone(&self) -> Result<Name, ProtoError> {
@@ -336,6 +504,10 @@ impl ZoneConfig {
             ZoneTypeConfig::Hint { .. } => ZoneType::Hint,
         }
     }
+}
+
+fn empty_stores_error<T>() -> Result<T, String> {
+    Result::Err("empty [[zones.stores]] in config".to_owned())
 }
 
 #[derive(Deserialize, Debug)]
@@ -458,6 +630,36 @@ impl ServerZoneConfig {
     pub fn keys(&self) -> &[dnssec::KeyConfig] {
         &self.keys
     }
+}
+
+#[cfg(feature = "dnssec-ring")]
+async fn load_keys(
+    authority: &mut impl DnssecAuthority<Lookup = impl Send + Sync + Sized + 'static>,
+    zone_name: Name,
+    server_config: &ServerZoneConfig,
+) -> Result<(), String> {
+    if server_config.is_dnssec_enabled() {
+        for key_config in server_config.keys() {
+            key_config.load(authority, &zone_name).await?;
+        }
+
+        info!("signing zone: {zone_name}");
+        authority
+            .secure_zone()
+            .await
+            .map_err(|err| format!("failed to sign zone {zone_name}: {err}"))?;
+    }
+    Ok(())
+}
+
+#[cfg(not(feature = "dnssec-ring"))]
+#[allow(clippy::unnecessary_wraps)]
+async fn load_keys<T>(
+    _authority: &mut T,
+    _zone_name: Name,
+    _server_config: &ServerZoneConfig,
+) -> Result<(), String> {
+    Ok(())
 }
 
 /// Enumeration over store types for secondary nameservers.
